+++ src/com/gildorymrp/gildorymclasses/EntityDeathListener.java
@@ -146,7 +146,7 @@ implements Listener
 				Race race = cardMap.get(player.getName()).getRace();
 				Integer pvpHealth = CharacterCard.calculateHealth(clazz, race, level);
 				
-				player.setExp(((Integer)experienceMap.get(player.getName())).intValue() / expToNextLevel);
+				player.setExp((float) (((Integer) experienceMap.get(player.getName())).intValue() / (double) expToNextLevel));
 				player.setLevel(((Integer)levelMap.get(player.getName())).intValue());
 				player.setMaxHealth(pvpHealth * 5);
 				Economy economy = (Economy)this.plugin.getServer().getServicesManager().getRegistration(Economy.class).getProvider();
