+++ src/com/gildorymrp/gildorymclasses/PlayerJoinListener.java
@@ -48,7 +48,7 @@ public class PlayerJoinListener implements Listener {
 		Race race = cardMap.get(player.getName()).getRace();
 		Integer pvpHealth = CharacterCard.calculateHealth(clazz, race, level);
 		
-		player.setExp(((Integer) experienceMap.get(player.getName())).intValue() / expToNextLevel);
+		player.setExp((float) (((Integer) experienceMap.get(player.getName())).intValue() / (double) expToNextLevel));
 		player.setMaxHealth(pvpHealth * 5);
 		player.setLevel(((Integer) levelMap.get(player.getName())).intValue());
 	}
